backends/sql_analysis/requirements.txt
fastapi
uvicorn
mysql-connector-python
pandas
google-generativeai
pydantic
python-dotenv
fpdf2